feat: inicio y cierre de sesión con clerk
