+++ pages/menu/index.jsx
@@ -6,14 +6,8 @@ import RandyTextRight from "@/components/RandyTextRight";
 import RandyTextLeft from "@/components/RandyTextLeft";
 import BackgroundModal from "@/components/BackgroundModal";
 import { useState, useEffect } from "react";
-import { useRouter } from "next/router";
-import { useUser } from "@clerk/nextjs";
-import { getCookieValueByName } from "@/components/utils/getCookieValueByName";
-
 export default function Menu() {
   const [background, setBackground] = useState(null);
-  const { isLoaded, user } = useUser([]);
-  const router = useRouter();
 
   const updateBackground = () => {
     const bgNew = localStorage.getItem("bg");
@@ -25,53 +19,7 @@ export default function Menu() {
   useEffect(() => {
     updateBackground();
   }, []);
-  //clerck
-  useEffect(() => {
-    if (isLoaded && user) {
-      const saveClerkUserDataOnLocalHost = async () => {
-        const response = await fetch("http://localhost:3005/users/email", {
-          method: "POST",
-          body: JSON.stringify({
-            email: user.emailAddresses[0].emailAddress,
-          }),
-          headers: {
-            "Content-type": "application/json; charset=UTF-8",
-          },
-        }).catch((error) => {
-          console.log("Error: ", error);
-        });
 
-        const data = await response?.json();
-        if (data) {
-          const cookieName = "__clerk_db_jwt";
-          const cookieValue = getCookieValueByName(cookieName);
-          const idUser = JSON.stringify(data?.data?._id);
-
-          if (cookieValue && data) {
-            localStorage.setItem("token", cookieValue);
-            localStorage.setItem("userID", idUser.replaceAll('"', ""));
-            localStorage.setItem("username", JSON.stringify(data.data.name));
-            localStorage.setItem("avatar", JSON.stringify(data.data.avatar));
-            localStorage.setItem(
-              "favs",
-              JSON.stringify(data.data.sandiasFavoritas)
-            );
-            localStorage.setItem(
-              "view",
-              JSON.stringify(data.data.sandiasVistas)
-            );
-            localStorage.setItem(
-              "achieve",
-              JSON.stringify(data.data.achievements)
-            );
-            localStorage.setItem("score", JSON.stringify(data.data.score));
-          }
-        }
-      };
-
-      saveClerkUserDataOnLocalHost();
-    }
-  }, [isLoaded, user, router]);
   return (
     <div className="w-full max-h-screen max-w-screen flex flex-col bg-white pb-8 overflow-hidden">
       <Navbar />
